# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AlfaUnipac GV — Sistema de Gestão Acadêmica API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /disciplinas:
    get:
      operationId: listDisciplinas
      tags: [disciplinas]
      summary: List all disciplines
      security:
        - bearerAuth: []
      parameters:
        - name: curso
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of disciplines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Disciplina"
    post:
      operationId: createDisciplina
      tags: [disciplinas]
      summary: Create a discipline
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDisciplinaBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disciplina"
  /disciplinas/{id}:
    patch:
      operationId: updateDisciplina
      tags: [disciplinas]
      summary: Update a discipline
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDisciplinaBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Disciplina"
    delete:
      operationId: deleteDisciplina
      tags: [disciplinas]
      summary: Delete a discipline
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # TURMAS
...
  /chamadas:
    get:
      operationId: listChamadas
      tags: [chamadas]
      summary: List attendance records
      security:
        - bearerAuth: []
      parameters:
        - name: turmaId
          in: query
          schema:
            type: integer
        - name: alunoId
          in: query
          schema:
            type: integer
        - name: data
          in: query
          schema:
            type: string
        - name: dataInicio
          in: query
          schema:
            type: string
        - name: dataFim
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of attendance records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Chamada"
    post:
      operationId: saveChamada
      tags: [chamadas]
      summary: Save attendance for a turma on a date
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveChamadaBody"
      responses:
        "200":
          description: Saved attendance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveChamadaResponse"
  /chamadas/frequencia/{alunoId}:
    get:
      operationId: getFrequenciaAluno
      tags: [chamadas]
      summary: Get attendance summary for a student across all turmas
      security:
        - bearerAuth: []
      parameters:
        - name: alunoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Frequency summary per turma
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FrequenciaResumo"
  /chamadas/datas/{turmaId}:
    get:
      operationId: getDatasRegistradas
      tags: [chamadas]
      summary: Get registered attendance dates for a turma
      security:
        - bearerAuth: []
      parameters:
        - name: turmaId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of dates
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  # RETENCAO
...
  /documentos:
    get:
      operationId: listDocumentos
      tags: [documentos]
      summary: List documents
      security:
        - bearerAuth: []
      parameters:
        - name: alunoId
          in: query
          schema:
            type: integer
        - name: tipo
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: curso
          in: query
          schema:
            type: string
        - name: dataInicio
          in: query
          schema:
            type: string
        - name: dataFim
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentoDetalhado"
    post:
      operationId: createDocumento
      tags: [documentos]
      summary: Register a document
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentoBody"
      responses:
        "201":
          description: Created document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Documento"
  /documentos/{id}:
    get:
      operationId: getDocumento
      tags: [documentos]
      summary: Get a document by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentoDetalhado"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateDocumentoStatus
      tags: [documentos]
      summary: Update document status (approve/reject)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentoStatusBody"
      responses:
        "200":
          description: Updated document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Documento"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /documentos/pendentes/count:
    get:
      operationId: getPendentesCount
      tags: [documentos]
      summary: Count pending documents
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Count
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CountResponse"

  # BAP
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    CountResponse:
      type: object
      properties:
        count:
          type: integer
      required:
        - count

    # Auth
...
    Disciplina:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        curso:
          type: string
        semestre:
          type: integer
        cargaHorariaTotal:
          type: integer
        professor:
          type: ["string", "null"]
      required:
        - id
        - nome
        - curso
        - semestre
        - cargaHorariaTotal
    CreateDisciplinaBody:
      type: object
      properties:
        nome:
          type: string
        curso:
          type: string
        semestre:
          type: integer
        cargaHorariaTotal:
          type: integer
        professor:
          type: ["string", "null"]
      required:
        - nome
        - curso
        - semestre
        - cargaHorariaTotal
    UpdateDisciplinaBody:
      type: object
      properties:
        nome:
          type: string
        curso:
          type: string
        semestre:
          type: integer
        cargaHorariaTotal:
          type: integer
        professor:
          type: ["string", "null"]

    # Turma
...
    Chamada:
      type: object
      properties:
        id:
          type: integer
        turmaId:
          type: integer
        alunoId:
          type: integer
        data:
          type: string
        presente:
          type: boolean
        justificada:
          type: boolean
        tipoJustificativa:
          type: ["string", "null"]
        observacao:
          type: ["string", "null"]
        alunoNome:
          type: ["string", "null"]
      required:
        - id
        - turmaId
        - alunoId
        - data
        - presente
        - justificada
    SaveChamadaBody:
      type: object
      properties:
        turmaId:
          type: integer
        data:
          type: string
        registros:
          type: array
          items:
            $ref: "#/components/schemas/ChamadaRegistro"
      required:
        - turmaId
        - data
        - registros
    ChamadaRegistro:
      type: object
      properties:
        alunoId:
          type: integer
        presente:
          type: boolean
        justificada:
          type: boolean
        tipoJustificativa:
          type: ["string", "null"]
        observacao:
          type: ["string", "null"]
      required:
        - alunoId
        - presente
        - justificada
    SaveChamadaResponse:
      type: object
      properties:
        saved:
          type: integer
        retencaoFlagged:
          type: integer
      required:
        - saved
        - retencaoFlagged
    FrequenciaResumo:
      type: object
      properties:
        turmaId:
          type: integer
        disciplinaNome:
          type: string
        periodo:
          type: string
        totalAulas:
          type: integer
        presencas:
          type: integer
        faltas:
          type: integer
        faltasJustificadas:
          type: integer
        percentualFaltas:
          type: number
        emRetencao:
          type: boolean
      required:
        - turmaId
        - disciplinaNome
        - periodo
        - totalAulas
        - presencas
        - faltas
        - faltasJustificadas
        - percentualFaltas
        - emRetencao

    # Retencao
...
    Documento:
      type: object
      properties:
        id:
          type: integer
        alunoId:
          type: integer
        tipo:
          type: string
          enum: [Atestado, Declaracao_Trabalho, Licenca_Maternidade]
        dataEntrega:
          type: string
        dataInicioPeriodo:
          type: string
        dataFimPeriodo:
          type: string
        arquivoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [Pendente, Aprovado, Rejeitado]
        observacao:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - alunoId
        - tipo
        - dataEntrega
        - dataInicioPeriodo
        - dataFimPeriodo
        - status
        - createdAt
    DocumentoDetalhado:
      type: object
      properties:
        id:
          type: integer
        alunoId:
          type: integer
        tipo:
          type: string
          enum: [Atestado, Declaracao_Trabalho, Licenca_Maternidade]
        dataEntrega:
          type: string
        dataInicioPeriodo:
          type: string
        dataFimPeriodo:
          type: string
        arquivoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [Pendente, Aprovado, Rejeitado]
        observacao:
          type: ["string", "null"]
        createdAt:
          type: string
        alunoNome:
          type: string
        alunoCurso:
          type: string
      required:
        - id
        - alunoId
        - tipo
        - dataEntrega
        - dataInicioPeriodo
        - dataFimPeriodo
        - status
        - createdAt
        - alunoNome
        - alunoCurso
    CreateDocumentoBody:
      type: object
      properties:
        alunoId:
          type: integer
        tipo:
          type: string
          enum: [Atestado, Declaracao_Trabalho, Licenca_Maternidade]
        dataEntrega:
          type: string
        dataInicioPeriodo:
          type: string
        dataFimPeriodo:
          type: string
        arquivoUrl:
          type: ["string", "null"]
        observacao:
          type: ["string", "null"]
      required:
        - alunoId
        - tipo
        - dataEntrega
        - dataInicioPeriodo
        - dataFimPeriodo
    UpdateDocumentoStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [Aprovado, Rejeitado]
        observacao:
          type: ["string", "null"]
      required:
        - status

    # BAP
...